Add other source of problems

+++ Debugging/CUnit/Fac/fac.c
@@ -1,9 +1,8 @@
 #include "fac.h"
 
-long fac(long n) {
-    long f = 1;
+int fac(int n) {
+    int f = 1;
     while (n > 1)
         f *= --n;
     return f;
 }
-
